move util functions into Gateway.ts
