+++ tests/utils.ts
@@ -29,34 +29,6 @@ export interface MsgFeeReport {
     importFee: bigint;
 }
 
-export function evmAddressToSlice(address: string): Slice {
-    if (address.length !== 42) {
-        throw new Error(`Invalid EVM address: ${address}`);
-    }
-
-    // Remove the '0x' prefix
-    const hexString = address.slice(2);
-
-    // Convert to Buffer
-    const buffer = Buffer.from(hexString, 'hex');
-    if (buffer.length !== 20) {
-        throw new Error(`Invalid Buffer length: ${buffer.length}`);
-    }
-
-    return beginCell().storeBuffer(buffer).asSlice();
-}
-
-// loads Slice to hex string `0x...`
-export function loadHexStringFromSlice(s: Slice, bytes: number): string {
-    return loadHexStringFromBuffer(s.loadBuffer(bytes));
-}
-
-export function loadHexStringFromBuffer(b: Buffer): string {
-    const hex = b.toString('hex');
-
-    return `0x${hex}`;
-}
-
 export function reportTXFees(tx: Transaction): TxFeeReport {
     const desc = tx.description as TransactionDescriptionGeneric;
 
@@ -130,29 +102,6 @@ function trimSuffix(str: string, suffix: string): string {
     }
 
     return str;
-}
-
-/**
- * Signs a cell with secp256k1 signature into a Slice (65 bytes)
- * @param signer
- * @param cell
- * @param log
- */
-export function signCellECDSA(signer: Wallet, cell: Cell, log: boolean = false): Slice {
-    const hash = cell.hash();
-    const sig = signer.signingKey.sign(hash);
-
-    // https://docs.ton.org/learn/tvm-instructions/instructions
-    //
-    // `ECRECOVER` Recovers public key from signature...
-    // Takes 32-byte hash as uint256 hash; 65-byte signature as uint8 v and uint256 r, s.
-    const [v, r, s] = [Number(sig.v), BigInt(sig.r), BigInt(sig.s)];
-
-    if (log) {
-        console.log('signCellECDSA', { v, r, s });
-    }
-
-    return beginCell().storeUint(v, 8).storeUint(r, 256).storeUint(s, 256).asSlice();
 }
 
 /**
